requirements.txt
fastmcp
requests
fastapi
uvicorn
python-dotenv

